pyproject.toml
[project]
name = "sonos-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0.1",
    "pigpio>=1.78",
    "python-dotenv>=1.1.1",
    "soco>=0.30.11",
]
